Refactor gallery vault screen and introduce vault settings screen
- Remove unused imports and methods from the gallery vault screen
- Replace settings sheet with navigation to a new vault settings screen
- Implement vault settings screen with options for security, appearance, and storage settings
- Add functionality for changing security settings and toggling encryption and secure delete features
