+++ lib/screens/gallery_vault_screen.dart
@@ -9,14 +9,11 @@ import 'package:path_provider/path_provider.dart';
 import '../models/vaulted_file.dart';
 import '../models/album.dart';
 import '../providers/vault_providers.dart';
-import '../services/auth_service.dart';
 import '../services/auto_kill_service.dart';
 import '../services/file_import_service.dart';
-import '../services/vault_service.dart';
 import '../themes/app_colors.dart';
 import '../utils/toast_utils.dart';
 import '../utils/responsive_utils.dart';
-import '../providers/theme_provider.dart';
 import '../widgets/office_conversion_confirm_dialog.dart';
 import 'albums_screen.dart';
 import 'favorites_screen.dart';
@@ -29,9 +26,7 @@ import 'media_picker_screen.dart';
 import 'document_picker_screen.dart';
 import 'package:photo_manager/photo_manager.dart' hide AlbumType;
 import 'camera_screen.dart';
-import 'local_backup_screen.dart';
-import 'change_security_screen.dart';
-import 'accent_color_picker_screen.dart';
+import 'vault_settings_screen.dart';
 import '../widgets/operation_progress_sheet.dart';
 
 /// Gallery vault screen - main screen after authentication
@@ -44,10 +39,6 @@ class GalleryVaultScreen extends ConsumerStatefulWidget {
 
 class _GalleryVaultScreenState extends ConsumerState<GalleryVaultScreen>
     with SingleTickerProviderStateMixin {
-  static const List<int> _lockoutAttemptOptions = [3, 5, 7, 10];
-  static const List<int> _lockoutDurationOptions = [30, 60, 300, 900];
-  static const List<int> _wipeAttemptOptions = [10, 15, 20, 30];
-
   late TabController _tabController;
   final FileImportService _importService = FileImportService.instance;
 
@@ -200,7 +191,7 @@ class _GalleryVaultScreenState extends ConsumerState<GalleryVaultScreen>
                 );
                 break;
               case 'settings':
-                _showSettingsSheet();
+                _openSettingsScreen();
                 break;
               case 'refresh':
                 ref.read(vaultNotifierProvider.notifier).refresh();
@@ -1603,7 +1594,7 @@ class _GalleryVaultScreenState extends ConsumerState<GalleryVaultScreen>
                         title: 'Security Settings',
                         onTap: () {
                           Navigator.pop(context);
-                          _showSettingsSheet();
+                          _openSettingsScreen();
                         },
                       ),
                       _buildDrawerItem(
@@ -2439,494 +2430,12 @@ class _GalleryVaultScreenState extends ConsumerState<GalleryVaultScreen>
     }
   }
 
-  void _showSettingsSheet() {
-    showModalBottomSheet(
-      context: context,
-      backgroundColor: Colors.transparent,
-      isScrollControlled: true,
-      builder: (context) => Container(
-        constraints: BoxConstraints(
-          maxHeight: MediaQuery.of(context).size.height * 0.85,
-        ),
-        decoration: BoxDecoration(
-          color: context.backgroundColor,
-          borderRadius: const BorderRadius.vertical(top: Radius.circular(24)),
-        ),
-        child: Consumer(
-          builder: (context, ref, _) {
-            final settingsAsync = ref.watch(vaultSettingsProvider);
-            return SingleChildScrollView(
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  Container(
-                    margin: const EdgeInsets.only(top: 12),
-                    width: 40,
-                    height: 4,
-                    decoration: BoxDecoration(
-                      color: context.borderColor,
-                      borderRadius: BorderRadius.circular(2),
-                    ),
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.all(20),
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          'Security Settings',
-                          style: TextStyle(
-                            fontSize: 20,
-                            fontWeight: FontWeight.bold,
-                            color: context.textPrimary,
-                            fontFamily: 'ProductSans',
-                          ),
-                        ),
-                        const SizedBox(height: 16),
-                        ListTile(
-                          leading: Icon(Icons.security_outlined,
-                              color: context.accentColor),
-                          title: const Text('Change Security',
-                              style: TextStyle(fontFamily: 'ProductSans')),
-                          subtitle: Text('Change password, PIN, or biometric',
-                              style: TextStyle(
-                                  fontFamily: 'ProductSans',
-                                  fontSize: 12,
-                                  color: context.textTertiary)),
-                          onTap: () {
-                            Navigator.pop(context);
-                            Navigator.push(
-                                context,
-                                MaterialPageRoute(
-                                    builder: (context) =>
-                                        const ChangeSecurityScreen()));
-                          },
-                          contentPadding: EdgeInsets.zero,
-                        ),
-                        ListTile(
-                          leading: Icon(Icons.backup_outlined,
-                              color: context.accentColor),
-                          title: const Text('Local backup',
-                              style: TextStyle(fontFamily: 'ProductSans')),
-                          subtitle: Text('Save vault as ZIP to a folder',
-                              style: TextStyle(
-                                  fontFamily: 'ProductSans',
-                                  fontSize: 12,
-                                  color: context.textTertiary)),
-                          onTap: () {
-                            Navigator.pop(context);
-                            Navigator.push(
-                                context,
-                                MaterialPageRoute(
-                                    builder: (context) =>
-                                        const LocalBackupScreen()));
-                          },
-                          contentPadding: EdgeInsets.zero,
-                        ),
-                        const SizedBox(height: 16),
-                        _buildThemeToggle(context, ref),
-                        const SizedBox(height: 8),
-                        _buildAccentColorOption(context, ref),
-                        const SizedBox(height: 16),
-                        settingsAsync.when(
-                          loading: () =>
-                              const Center(child: CircularProgressIndicator()),
-                          error: (_, __) =>
-                              const Text('Failed to load settings'),
-                          data: (settings) => Column(
-                            children: [
-                              SwitchListTile(
-                                title: const Text('Encrypt New Files',
-                                    style:
-                                        TextStyle(fontFamily: 'ProductSans')),
-                                subtitle: Text(
-                                    'AES-256 encryption for all new imports',
-                                    style: TextStyle(
-                                        fontFamily: 'ProductSans',
-                                        fontSize: 12,
-                                        color: context.textTertiary)),
-                                value: settings.encryptionEnabled,
-                                onChanged: (value) async {
-                                  await _saveVaultSettings(
-                                    settings.copyWith(
-                                      encryptionEnabled: value,
-                                    ),
-                                  );
-                                },
-                                activeThumbColor: context.accentColor,
-                                contentPadding: EdgeInsets.zero,
-                              ),
-                              SwitchListTile(
-                                title: const Text('Secure Delete',
-                                    style:
-                                        TextStyle(fontFamily: 'ProductSans')),
-                                subtitle: Text(
-                                    'Overwrite files before deletion',
-                                    style: TextStyle(
-                                        fontFamily: 'ProductSans',
-                                        fontSize: 12,
-                                        color: context.textTertiary)),
-                                value: settings.secureDelete,
-                                onChanged: (value) async {
-                                  await _saveVaultSettings(
-                                    settings.copyWith(
-                                      secureDelete: value,
-                                    ),
-                                  );
-                                },
-                                activeThumbColor: context.accentColor,
-                                contentPadding: EdgeInsets.zero,
-                              ),
-                              SwitchListTile(
-                                title: const Text('Compress Media',
-                                    style:
-                                        TextStyle(fontFamily: 'ProductSans')),
-                                subtitle: Text(
-                                    'Reduce file size for images and videos',
-                                    style: TextStyle(
-                                        fontFamily: 'ProductSans',
-                                        fontSize: 12,
-                                        color: context.textTertiary)),
-                                value: settings.compressionEnabled,
-                                onChanged: (value) async {
-                                  await _saveVaultSettings(
-                                    settings.copyWith(
-                                      compressionEnabled: value,
-                                    ),
-                                  );
-                                },
-                                activeThumbColor: context.accentColor,
-                                contentPadding: EdgeInsets.zero,
-                              ),
-                              const SizedBox(height: 8),
-                              Divider(color: context.borderColor),
-                              const SizedBox(height: 8),
-                              SwitchListTile(
-                                title: const Text(
-                                  'Failed Unlock Protection',
-                                  style: TextStyle(fontFamily: 'ProductSans'),
-                                ),
-                                subtitle: Text(
-                                  'Adds a cooldown after repeated wrong PIN, password, or biometric attempts',
-                                  style: TextStyle(
-                                    fontFamily: 'ProductSans',
-                                    fontSize: 12,
-                                    color: context.textTertiary,
-                                  ),
-                                ),
-                                value: settings.failedUnlockProtectionEnabled,
-                                onChanged: (value) async {
-                                  await _toggleFailedUnlockProtection(
-                                    settings,
-                                    value,
-                                  );
-                                },
-                                activeThumbColor: context.accentColor,
-                                contentPadding: EdgeInsets.zero,
-                              ),
-                              if (settings.failedUnlockProtectionEnabled) ...[
-                                _buildSecurityOptionDropdown(
-                                  title: 'Attempts Before Cooldown',
-                                  subtitle:
-                                      'How many failed unlocks are allowed before access is paused',
-                                  value:
-                                      settings.maxFailedAttemptsBeforeLockout,
-                                  options: _lockoutAttemptOptions,
-                                  labelBuilder: (value) => value.toString(),
-                                  onChanged: (value) async {
-                                    if (value == null) return;
-                                    await _saveUnlockProtectionSettings(
-                                      settings.copyWith(
-                                        maxFailedAttemptsBeforeLockout: value,
-                                      ),
-                                    );
-                                  },
-                                ),
-                                _buildSecurityOptionDropdown(
-                                  title: 'Cooldown Timer',
-                                  subtitle:
-                                      'How long unlock stays disabled after the cooldown threshold is hit',
-                                  value: settings.lockoutDurationSeconds,
-                                  options: _lockoutDurationOptions,
-                                  labelBuilder: _lockoutDurationLabel,
-                                  onChanged: (value) async {
-                                    if (value == null) return;
-                                    await _saveUnlockProtectionSettings(
-                                      settings.copyWith(
-                                        lockoutDurationSeconds: value,
-                                      ),
-                                    );
-                                  },
-                                ),
-                                SwitchListTile(
-                                  title: const Text(
-                                    'Wipe Vault At Hard Limit',
-                                    style: TextStyle(fontFamily: 'ProductSans'),
-                                  ),
-                                  subtitle: Text(
-                                    'Permanently erases real and decoy vault files after too many failed unlocks',
-                                    style: TextStyle(
-                                      fontFamily: 'ProductSans',
-                                      fontSize: 12,
-                                      color: context.textTertiary,
-                                    ),
-                                  ),
-                                  value: settings.wipeVaultOnMaxFailedAttempts,
-                                  onChanged: (value) async {
-                                    await _toggleVaultWipeProtection(
-                                      settings,
-                                      value,
-                                    );
-                                  },
-                                  activeThumbColor: AppColors.error,
-                                  contentPadding: EdgeInsets.zero,
-                                ),
-                                if (settings.wipeVaultOnMaxFailedAttempts) ...[
-                                  _buildSecurityOptionDropdown(
-                                    title: 'Attempts Before Wipe',
-                                    subtitle:
-                                        'Locker erases all vault files when this failed-attempt total is reached',
-                                    value: settings.maxFailedAttemptsBeforeWipe,
-                                    options: _wipeAttemptOptions,
-                                    labelBuilder: (value) => value.toString(),
-                                    onChanged: (value) async {
-                                      if (value == null) return;
-                                      await _saveUnlockProtectionSettings(
-                                        settings.copyWith(
-                                          maxFailedAttemptsBeforeWipe: value,
-                                        ),
-                                      );
-                                    },
-                                  ),
-                                  Container(
-                                    width: double.infinity,
-                                    margin: const EdgeInsets.only(top: 8),
-                                    padding: const EdgeInsets.all(12),
-                                    decoration: BoxDecoration(
-                                      color: AppColors.error
-                                          .withValues(alpha: 0.08),
-                                      borderRadius: BorderRadius.circular(12),
-                                      border: Border.all(
-                                        color: AppColors.error
-                                            .withValues(alpha: 0.18),
-                                      ),
-                                    ),
-                                    child: Text(
-                                      'Warning: wiping the vault is permanent and cannot be undone.',
-                                      style: TextStyle(
-                                        fontFamily: 'ProductSans',
-                                        fontSize: 12,
-                                        color: AppColors.error,
-                                      ),
-                                    ),
-                                  ),
-                                ],
-                              ],
-                            ],
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                  SizedBox(height: MediaQuery.of(context).padding.bottom),
-                ],
-              ),
-            );
-          },
-        ),
+  void _openSettingsScreen() {
+    Navigator.push(
+      context,
+      MaterialPageRoute(
+        builder: (context) => const VaultSettingsScreen(),
       ),
-    );
-  }
-
-  Future<void> _saveVaultSettings(VaultSettings settings) async {
-    await ref.read(vaultServiceProvider).updateSettings(settings);
-    ref.invalidate(vaultSettingsProvider);
-  }
-
-  Future<void> _saveUnlockProtectionSettings(VaultSettings settings) async {
-    await _saveVaultSettings(settings);
-    await AuthService().resetUnlockAttempts();
-  }
-
-  Future<void> _toggleFailedUnlockProtection(
-    VaultSettings settings,
-    bool enabled,
-  ) async {
-    await _saveUnlockProtectionSettings(
-      settings.copyWith(failedUnlockProtectionEnabled: enabled),
-    );
-  }
-
-  Future<void> _toggleVaultWipeProtection(
-    VaultSettings settings,
-    bool enabled,
-  ) async {
-    if (enabled) {
-      final confirmed = await _confirmDangerousWipeProtection();
-      if (confirmed != true) return;
-    }
-
-    await _saveUnlockProtectionSettings(
-      settings.copyWith(wipeVaultOnMaxFailedAttempts: enabled),
-    );
-  }
-
-  Future<bool?> _confirmDangerousWipeProtection() {
-    return showDialog<bool>(
-      context: context,
-      builder: (dialogContext) => AlertDialog(
-        backgroundColor: Theme.of(dialogContext).scaffoldBackgroundColor,
-        title: const Text(
-          'Enable Vault Wipe?',
-          style: TextStyle(fontFamily: 'ProductSans'),
-        ),
-        content: const Text(
-          'When the failed-attempt limit is reached, Locker will permanently erase the real and decoy vault files on this device. This cannot be undone.',
-          style: TextStyle(fontFamily: 'ProductSans'),
-        ),
-        actions: [
-          TextButton(
-            onPressed: () => Navigator.pop(dialogContext, false),
-            child: const Text('Cancel'),
-          ),
-          FilledButton(
-            onPressed: () => Navigator.pop(dialogContext, true),
-            child: const Text('Enable'),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildSecurityOptionDropdown({
-    required String title,
-    required String subtitle,
-    required int value,
-    required List<int> options,
-    required String Function(int value) labelBuilder,
-    required ValueChanged<int?> onChanged,
-  }) {
-    return ListTile(
-      contentPadding: EdgeInsets.zero,
-      title: Text(
-        title,
-        style: const TextStyle(fontFamily: 'ProductSans'),
-      ),
-      subtitle: Text(
-        subtitle,
-        style: TextStyle(
-          fontFamily: 'ProductSans',
-          fontSize: 12,
-          color: context.textTertiary,
-        ),
-      ),
-      trailing: DropdownButtonHideUnderline(
-        child: DropdownButton<int>(
-          value: value,
-          borderRadius: BorderRadius.circular(12),
-          items: options
-              .map(
-                (option) => DropdownMenuItem<int>(
-                  value: option,
-                  child: Text(
-                    labelBuilder(option),
-                    style: const TextStyle(fontFamily: 'ProductSans'),
-                  ),
-                ),
-              )
-              .toList(),
-          onChanged: onChanged,
-        ),
-      ),
-    );
-  }
-
-  String _lockoutDurationLabel(int seconds) {
-    if (seconds < 60) {
-      return '${seconds}s';
-    }
-
-    final minutes = seconds ~/ 60;
-    return minutes == 1 ? '1 min' : '$minutes min';
-  }
-
-  Widget _buildThemeToggle(BuildContext context, WidgetRef ref) {
-    final isDarkMode = ref.watch(isDarkModeProvider);
-
-    return SwitchListTile(
-      title: const Text(
-        'Dark Mode',
-        style: TextStyle(fontFamily: 'ProductSans'),
-      ),
-      subtitle: Text(
-        isDarkMode ? 'Eye-friendly dark theme' : 'Clean light theme',
-        style: TextStyle(
-          fontFamily: 'ProductSans',
-          fontSize: 12,
-          color: context.textTertiary,
-        ),
-      ),
-      secondary: Icon(
-        isDarkMode ? Icons.dark_mode : Icons.light_mode,
-        color: context.accentColor,
-      ),
-      value: isDarkMode,
-      onChanged: (value) {
-        ref.read(themeModeProvider.notifier).toggleTheme();
-      },
-      activeThumbColor: context.accentColor,
-      contentPadding: EdgeInsets.zero,
-    );
-  }
-
-  Widget _buildAccentColorOption(BuildContext context, WidgetRef ref) {
-    final accentColor = ref.watch(accentColorProvider);
-
-    return ListTile(
-      leading: Container(
-        width: 32,
-        height: 32,
-        decoration: BoxDecoration(
-          gradient: LinearGradient(
-            colors: [
-              accentColor.getColor(
-                  context.isDarkMode ? Brightness.dark : Brightness.light),
-              accentColor.getVariantColor(
-                  context.isDarkMode ? Brightness.dark : Brightness.light),
-            ],
-          ),
-          borderRadius: BorderRadius.circular(8),
-          border: Border.all(
-            color: context.isDarkMode
-                ? Colors.white.withValues(alpha: 0.2)
-                : Colors.black.withValues(alpha: 0.1),
-            width: 1.5,
-          ),
-        ),
-      ),
-      title: const Text(
-        'Accent Color',
-        style: TextStyle(fontFamily: 'ProductSans'),
-      ),
-      subtitle: Text(
-        accentColor.name,
-        style: TextStyle(
-          fontFamily: 'ProductSans',
-          fontSize: 12,
-          color: context.textTertiary,
-        ),
-      ),
-      trailing: const Icon(Icons.chevron_right),
-      onTap: () {
-        Navigator.pop(context);
-        Navigator.push(
-          context,
-          MaterialPageRoute(
-            builder: (context) => const AccentColorPickerScreen(),
-          ),
-        );
-      },
-      contentPadding: EdgeInsets.zero,
     );
   }
